Split MatchDetails into components; UI/spacing fixes
