+++ frontend/src/pages/matchDetailsPage/styles/matchHeader.css
@@ -1,0 +1,37 @@
+/* Actions row */
+.header-actions{display:flex;align-items:center;justify-content:space-between;gap:1rem;margin-bottom:1rem}
+.back-button{display:inline-flex;align-items:center;gap:.5rem;font-weight:800;font-size:.92rem;text-decoration:none;padding:.55rem 1rem;border-radius:10px;background:#1d2432;color:var(--text-200);border:1px solid rgba(209,213,219,.18);transition:all .18s}
+.back-button:hover{background:#232a3a;color:var(--text-100);border-color:rgba(209,213,219,.4);transform:translateY(-1px)}
+.right-actions{display:flex;align-items:center;gap:.5rem}
+.ghost-button{display:inline-flex;align-items:center;gap:.4rem;background:rgba(255,255,255,.06);border:1px solid var(--line);color:var(--text-200);padding:.45rem .8rem;border-radius:10px;text-decoration:none;font-weight:800}
+.ghost-button:hover{background:rgba(255,255,255,.09);color:var(--text-100)}
+.comp-pill{padding:.22rem .6rem;border-radius:999px;background:rgba(59,130,246,.18);border:1px solid rgba(59,130,246,.28);color:#dbeafe;font-weight:700;font-size:.8rem}
+
+/* Header (team/score box) */
+.match-header{ --logo:48px; --name-size:1.18rem; --score-size:2.1rem; padding:1rem var(--page-xpad); }
+.teams-vs{display:grid;grid-template-columns:minmax(0,1fr) minmax(70px,max-content) minmax(0,1fr);align-items:center;gap:1rem}
+.team{display:inline-flex;align-items:center;gap:.6rem;min-width:0}
+.team.right{justify-content:end}
+.logo{width:var(--logo);height:var(--logo);object-fit:contain;background:rgba(255,255,255,.06);border:1px solid rgba(255,255,255,.12);border-radius:8px;padding:.32rem}
+.name{font-size:var(--name-size);font-weight:800;color:var(--text-100);white-space:nowrap;overflow:hidden;text-overflow:ellipsis;letter-spacing:.01em;min-width:0}
+.score-block{display:flex;flex-direction:column;align-items:center;gap:.35rem}
+.score{font-size:var(--score-size);font-weight:900;letter-spacing:.02em;color:var(--text-100)}
+.status-chip{display:inline-flex;align-items:center;gap:.4rem;font-size:.86rem;padding:.12rem .7rem;border-radius:999px;background:rgba(255,255,255,.06);border:1px solid var(--line);color:var(--text-200);font-weight:800}
+.status-chip.live{background:rgba(248,113,113,.18);border-color:rgba(248,113,113,.35);color:#fecaca}
+.pulse{width:8px;height:8px;border-radius:50%;background:currentColor;animation:pulse 1.2s infinite ease-in-out}
+@keyframes pulse{0%,100%{opacity:.5;transform:scale(.9)}50%{opacity:1;transform:scale(1.2)}}
+.meta-row{margin-top:.7rem;display:flex;gap:.7rem;align-items:center;flex-wrap:wrap;opacity:.95;font-size:.96rem;color:var(--text-200)}
+.dot{opacity:.55}
+
+@media (max-width:680px){
+  .match-header{--logo:42px;--name-size:1.02rem;--score-size:1.9rem}
+  .status-chip{font-size:.8rem;padding:.1rem .55rem}
+}
+@media (max-width:540px){
+  .match-header{--logo:34px;--name-size:clamp(.86rem,3.6vw,1rem);--score-size:clamp(1.45rem,6.2vw,1.7rem)}
+  .teams-vs{gap:.7rem}
+  .meta-row{font-size:.88rem}
+}
+@media (max-width:380px){
+  .match-header{--logo:30px;--name-size:clamp(.82rem,3.9vw,.95rem);--score-size:clamp(1.35rem,6.5vw,1.55rem)}
+}
